Release/v0.0.3

1. Fix bug: BFOP 321, 386, 388, 387, 403, 406, 409, 415, 418, 408, 211, 388, 211, 267, 268, 326, 327, 334, 337
2. Add verfify_password column when update user password
3. Add cmd/cmd-class API for UI new page
4. Modify some UI style with style guideline
5. Fix some issue in display statistic chart

See merge request !3

+++ src/components/CategoryCard.vue
@@ -1,13 +1,13 @@
 
 <template>
   <div id="card-category">
-    <div id="card-category-header">
+    <div id="card-category-header" v-if="!selectOnly">
       <div id="card-category-header-block">
         <div id="card-category-row">
-          <div class="card-category-title">{{ $t('wordbank.categories') }}</div>
+          <div class="card-category-title">{{ $t('category.title') }}</div>
           <div v-if="canSetting" class="card-category-setting" @click="toggleEditMode">
-            <span v-if="isEditMode"> {{ $t('wordbank.leave_setting') }}</span>
-            <span v-else> {{ $t('wordbank.setting') }} </span>
+            <span v-if="isEditMode"> {{ $t('category.leave_setting') }}</span>
+            <span v-else> {{ $t('category.setting') }} </span>
           </div>
         </div>
         <search-input ref="categorySearchBox" v-model="categoryNameKeyword"></search-input>
@@ -15,18 +15,21 @@
       <div id="card-category-add-root" v-if="isEditMode && canCreate" @click="addRootCategory"> 
         <input id="add-root" type="text"
           v-if="isAddingRoot" v-model="rootName" ref="rootName" 
-          :placeholder="$t('wordbank.placeholder_category_name')"
+          :placeholder="$t('category.placeholder_category_name')"
           @compositionstart="setCompositionState(true)"
           @compositionend="setCompositionState(false)"
           @blur="confirmRootName"
           @keydown.enter="detectCompositionState"
           @keyup.enter="confirmRootName">
-        <span v-else>{{ $t('wordbank.add_rootcategory') }}</span>
+        <div v-else id="add-root-btn">
+          <icon icon-type="category_add" :size=16></icon>
+          <span> {{ $t('category.add_root') }} </span>
+        </div>
       </div>
     </div>
     <div id="card-category-content" ref="cardCategoryContent">
       <div v-if="!hasSearchResult" id="no-category-search-result">
-        {{ $t('wordbank.empty_category_search_result') }}
+        {{ $t('category.empty_search_result') }}
       </div>
       <category-tree-item v-for="(child, idx) in categoryTree.children" 
         v-if="child.visible"
@@ -40,17 +43,17 @@
         @itemNameChange="handleItemNameChange"
         @setActiveToAll="handleSetActiveToAll"></category-tree-item>
     </div>
-    <div id="card-category-footer" v-if="isEditMode">
+    <div id="card-category-footer" v-if="isEditMode && !selectOnly">
       <div v-if="allowSubCategory && canCreate" 
         class="card-category-setting-option option-addsub"
         :class="{'option-disabled': this.currentActiveItem.layer === maxLayer || this.currentActiveItem.cid < 0}"
         @click="addSubCategory">
-        {{ $t('wordbank.add_subcategory') }}
+        {{ $t('category.add_subcategory') }}
       </div>
       <div v-if="canDelete" class="card-category-setting-option option-delete"
         :class="{'option-disabled': !this.currentActiveItem.deletable}"
         @click="popDeleteCategory">
-        {{ $t('wordbank.delete_category') }}
+        {{ $t('category.delete_category') }}
       </div>
     </div>
   </div>
@@ -89,6 +92,10 @@ export default {
       default: false,
     },
     canCreate: {
+      type: Boolean,
+      default: false,
+    },
+    selectOnly: {
       type: Boolean,
       default: false,
     },
@@ -194,7 +201,9 @@ export default {
       this.currentActiveItem.route = route;
     },
     resetActiveItem(treeItem) {
-      treeItem.isActive = false;
+      if (treeItem.cid !== -1) {
+        treeItem.isActive = false;
+      }
       if (treeItem.children && treeItem.children.length > 0) {
         treeItem.children.forEach((child) => {
           this.resetActiveItem(child);
@@ -264,6 +273,9 @@ export default {
     },
     findItemParentRef(item) {
       const parent = this.findItemParent(this.categoryTree, item.cid);
+      if (parent.cid === -1) {
+        return undefined;
+      }
       const parentRefName = `${parent.cid}-${parent.name}`;
       return this.findItemRef(this, parentRefName);
     },
@@ -318,7 +330,12 @@ export default {
       if (this.isAddingRoot) {
         if (success) {
           this.resetActiveItem(this.categoryTree);
-          this.categoryTree.children.push(-1, 0, category);
+          this.categoryTree.children.push(category);
+          this.setActiveItem(this.categoryTree, category.cid);
+          this.$nextTick(() => {
+            const categoryContentBlock = this.$refs.cardCategoryContent;
+            categoryContentBlock.scrollTop = categoryContentBlock.scrollHeight;
+          });
         }
         this.rootName = '';
         this.isAddingRoot = false;
@@ -344,15 +361,25 @@ export default {
       };
       this.$popCheck(option);
     },
-    confirmDeleteCategory() { // v
+    confirmDeleteCategory() {
       this.$emit('deleteCategory', this.currentActiveItem);
     },
-    deleteCategorySuccess(success) {  // v
-      const parentRef = this.findItemParentRef(this.currentActiveItem);
-      parentRef.deleteCategorySuccess(success, this.currentActiveItem.cid);
+    deleteCategorySuccess(success) {
+      if (success) {
+        const parentRef = this.findItemParentRef(this.currentActiveItem);
+        if (parentRef === undefined) {  // parent is root
+          const idxToDel = this.categoryTree.children
+            .findIndex(child => child.id === this.currentActiveItem.cid);
+          this.categoryTree.children.splice(idxToDel);
+          // back to all
+          this.handleSetActiveToAll();
+        } else {
+          parentRef.deleteCategorySuccess(success, this.currentActiveItem.cid);
+        }
+      }
     },
 
-    toggleEditMode() {  // v
+    toggleEditMode() {
       this.isEditMode = !this.isEditMode;
       if (!this.isEditMode) {
         this.categoryNameKeyword = '';
@@ -378,6 +405,7 @@ export default {
 </script>
 <style lang="scss" scoped>
 @import 'styles/variable';
+$category-item-height: 32px;
 
 #card-category {
   display: flex;
@@ -411,13 +439,14 @@ export default {
 
     #card-category-add-root {
       background: #fcfcfc;
-      height: 36px;
-      line-height: 20px;
-      padding: 8px;
-      padding-left: 23px;
+      height: $category-item-height;
+      line-height: 18px;
+      padding: 10px;
       color: $color-primary;
       cursor: pointer;
       border-bottom: 1px solid $color-borderline;
+      display: flex;
+      align-items: center;
       input[type=text] {
         &#add-root {
           background: #fcfcfc;
@@ -431,14 +460,21 @@ export default {
           }   
         }
       }
+      #add-root-btn {
+        display: flex;
+        align-items: center;
+        span {
+          padding: 0 8px;
+        }
+      }
     }
   }
   #card-category-content {
     flex: 1 1 auto;
     overflow-y: auto;
     #no-category-search-result {
-      height: 36px;
-      line-height: 36px;
+      height: $category-item-height;
+      line-height: $category-item-height;
       text-align: center;
     }
   }
